Created a file for complete status operations
